Create main.go
A load balancer is a program thats responsable for keeping track of server loads, ie. which server
has the most amount of traffic, which is available right now, and which server to send traffic to.

We will use round robin method [or logic if u wanna call it that] to decide where to send traffic, its the simplest method to route traffic between servers. there are way more complex algorithms but
we will not be doing that.

Load balancers often have Reverse Proxy support
in our program we will have a client (could be a server acting as a client, jsut anything making a request to a server). Request will be  sent to the Reverse Proxy, were copying code for this cuz we are not that deep yet <3 as of 4/27/23. will look into it later but we got bigger fish to fry rn

FOR NOW THO we will use the http util package to create a reverse proxy server to redirect our request and hide the location from the server from the client.